Implement adding card 2
