+++ package.json
@@ -16,6 +16,7 @@
     "node-sass": "^4.14.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-redux": "^7.2.6",
     "react-scripts": "4.0.3",
     "react-show-more-button": "^1.0.0",
     "react-show-more-text": "^1.5.2",
