+++ src/components/Reducer.js
@@ -1,0 +1,12 @@
+
+
+const Reducer =(cards=[],action)=>{
+    if(action.type==="ADDCard"){
+        return [...cards, action.payload];
+    }
+
+    return cards;
+};
+   
+export default Reducer;
+
