+++ src/sass.scss
@@ -38,4 +38,26 @@ body {
 .firstRow{
   display: flex;
   justify-content: space-between;
+  border-radius: 20px;
+  // background-color: rgb(54, 166, 231);
+  // border: rgb(54, 166, 231);
 }
+
+.descRow{
+  display: flex;
+  justify-content: space-between;
+  border-radius: 20px;
+  background-color: rgb(54, 166, 231);
+  border: rgb(54, 166, 231);
+
+}
+
+// .Button{
+//   background-color: rgb(28, 103, 243);
+//   border-radius: 20px;
+//   border-color: blue;
+//   margin: 1;
+  
+
+
+//}
